Move Branch into Control.Effect.NonDet.
