+++ src/Control/Effect/Cut.hs
@@ -1,12 +1,9 @@
-{-# LANGUAGE DeriveTraversable, ExistentialQuantification, FlexibleContexts, FlexibleInstances, LambdaCase, MultiParamTypeClasses, StandaloneDeriving, TypeOperators, UndecidableInstances #-}
+{-# LANGUAGE DeriveFunctor, ExistentialQuantification, FlexibleContexts, FlexibleInstances, LambdaCase, MultiParamTypeClasses, StandaloneDeriving, TypeOperators, UndecidableInstances #-}
 module Control.Effect.Cut
 ( Cut(..)
 , cutfail
 , call
 , cut
-, Branch(..)
-, branch
-, runBranch
 , runCut
 , CutC(..)
 ) where
@@ -61,31 +58,6 @@ cut = pure () <|> cutfail
 {-# INLINE cut #-}
 
 
--- | The result of a nondeterministic branch of a computation.
-data Branch m e a
-  = None e
-  | Pure a
-  | Alt (m a) (m a)
-  deriving (Eq, Foldable, Functor, Ord, Show, Traversable)
-
--- | Case analysis for 'Branch', taking a value to use for 'Cut', a value to use for 'None', and a function to apply to the contents of 'Pure'.
---
---   prop> branch None Pure Alt a == (a :: Branch e [] a)
---   prop> branch (applyFun f) (applyFun g) (applyFun2 h) (None a :: Branch [] a) == applyFun f a
---   prop> branch (applyFun f) (applyFun g) (applyFun2 h) (Pure a :: Branch [] a) == applyFun g a
---   prop> branch (applyFun f) (applyFun g) (applyFun2 h) (Alt a b :: Branch [] a) == applyFun2 h a b
-branch :: (e -> a) -> (b -> a) -> (m b -> m b -> a) -> Branch m e b -> a
-branch f _ _ (None a)  = f a
-branch _ f _ (Pure a)  = f a
-branch _ _ f (Alt a b) = f a b
-{-# INLINE branch #-}
-
--- | Interpret a 'Branch' into an underlying 'Alternative' context.
-runBranch :: Alternative m => (e -> m a) -> Branch m e a -> m a
-runBranch f = branch f pure (<|>)
-{-# INLINE runBranch #-}
-
-
 -- | Run a 'Cut' effect within an underlying 'Alternative' instance (typically 'Eff' carrying a 'NonDet' effect).
 --
 --   prop> run (runNonDetOnce (runCut (pure a))) == Just a
@@ -115,5 +87,3 @@ instance (Alternative m, Carrier sig m, Effect sig, Monad m) => Carrier (Cut :+:
 -- >>> :seti -XFlexibleContexts
 -- >>> import Test.QuickCheck
 -- >>> import Control.Effect.Void
--- >>> instance (Arbitrary1 m, Arbitrary e) => Arbitrary1 (Branch m e) where liftArbitrary arb = frequency [(1, None <$> arbitrary), (3, Pure <$> arb), (3, Alt <$> liftArbitrary arb <*> liftArbitrary arb)]
--- >>> instance (Arbitrary1 m, Arbitrary e, Arbitrary a) => Arbitrary (Branch m e a) where arbitrary = arbitrary1
